api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic>=0.40.0
mistralai>=1.0.0
httpx>=0.27.0
pydantic>=2.8.0
python-dotenv==1.0.1
tqdm==4.66.6
scikit-learn>=1.5.0
numpy>=1.26.0
rank-bm25>=0.2.2
duckduckgo-search>=6.3.7
aiofiles>=24.1.0
